+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Show default agent in dashboard sidebar
 - Add AgentAvatar and AgentIdentity entity primitives
 - Add dedicated Feature Promos page to dev playground